Refactors App to track selectedProduct as an object instead of just an id
